+++ lib/PhpParser/Lexer/Emulative.php
@@ -4,27 +4,54 @@ namespace PhpParser\Lexer;
 
 use PhpParser\Error;
 use PhpParser\ErrorHandler;
+use PhpParser\Parser;
 
 class Emulative extends \PhpParser\Lexer
 {
     const PHP_7_3 = '7.3.0dev';
+    const PHP_7_4 = '7.4.0dev';
+
+    const FLEXIBLE_DOC_STRING_REGEX = <<<'REGEX'
+/<<<[ \t]*(['"]?)([a-zA-Z_\x80-\xff][a-zA-Z0-9_\x80-\xff]*)\1\r?\n
+(?:.*\r?\n)*?
+(?<indentation>\h*)\2(?![a-zA-Z_\x80-\xff])(?<separator>(?:;?[\r\n])?)/x
+REGEX;
+
+    const T_COALESCE_EQUAL = 1007;
 
     /**
-     * @var array Patches used to reverse changes introduced in the code
+     * @var mixed[] Patches used to reverse changes introduced in the code
      */
-    private $patches;
+    private $patches = [];
+
+    /**
+     * @param mixed[] $options
+     */
+    public function __construct(array $options = [])
+    {
+        parent::__construct($options);
+
+        // add emulated tokens here
+        $this->tokenMap[self::T_COALESCE_EQUAL] = Parser\Tokens::T_COALESCE_EQUAL;
+    }
 
     public function startLexing(string $code, ErrorHandler $errorHandler = null) {
         $this->patches = [];
-        $preparedCode = $this->prepareCode($code);
-        if (null === $preparedCode) {
+
+        if ($this->isEmulationNeeded($code) === false) {
             // Nothing to emulate, yay
             parent::startLexing($code, $errorHandler);
             return;
         }
 
         $collector = new ErrorHandler\Collecting();
+
+        // 1. emulation of heredoc and nowdoc new syntax
+        $preparedCode = $this->processHeredocNowdoc($code);
         parent::startLexing($preparedCode, $collector);
+
+        // 2. emulation of ??= token
+        $this->processCoaleseEqual($code);
         $this->fixupTokens();
 
         $errors = $collector->getErrors();
@@ -36,30 +63,60 @@ class Emulative extends \PhpParser\Lexer
         }
     }
 
-    /**
-     * Prepares code for emulation. If nothing has to be emulated null is returned.
-     *
-     * @param string $code
-     * @return null|string
-     */
-    private function prepareCode(string $code) {
+    private function isCoalesceEqualEmulationNeeded(string $code): bool
+    {
+        // skip version where this works without emulation
+        if (version_compare(\PHP_VERSION, self::PHP_7_4, '>=')) {
+            return false;
+        }
+
+        return strpos($code, '??=') !== false;
+    }
+
+    private function processCoaleseEqual(string $code)
+    {
+        if ($this->isCoalesceEqualEmulationNeeded($code) === false) {
+            return;
+        }
+
+        // We need to manually iterate and manage a count because we'll change
+        // the tokens array on the way
+        $line = 1;
+        for ($i = 0, $c = count($this->tokens); $i < $c; ++$i) {
+            if (isset($this->tokens[$i + 1])) {
+                if ($this->tokens[$i][0] === T_COALESCE && $this->tokens[$i + 1] === '=') {
+                    array_splice($this->tokens, $i, 2, [
+                        [self::T_COALESCE_EQUAL, '??=', $line]
+                    ]);
+                    $c--;
+                    continue;
+                }
+            }
+            if (\is_array($this->tokens[$i])) {
+                $line += substr_count($this->tokens[$i][1], "\n");
+            }
+        }
+    }
+
+    private function isHeredocNowdocEmulationNeeded(string $code): bool
+    {
+        // skip version where this works without emulation
         if (version_compare(\PHP_VERSION, self::PHP_7_3, '>=')) {
-            return null;
+            return false;
         }
 
-        if (strpos($code, '<<<') === false) {
-            // Definitely doesn't contain heredoc/nowdoc
-            return null;
+        return strpos($code, '<<<') !== false;
+    }
+
+    private function processHeredocNowdoc(string $code): string
+    {
+        if ($this->isHeredocNowdocEmulationNeeded($code) === false) {
+            return $code;
         }
 
-        $flexibleDocStringRegex = <<<'REGEX'
-/<<<[ \t]*(['"]?)([a-zA-Z_\x80-\xff][a-zA-Z0-9_\x80-\xff]*)\1\r?\n
-(?:.*\r?\n)*?
-(?<indentation>\h*)\2(?![a-zA-Z_\x80-\xff])(?<separator>(?:;?[\r\n])?)/x
-REGEX;
-        if (!preg_match_all($flexibleDocStringRegex, $code, $matches, PREG_SET_ORDER|PREG_OFFSET_CAPTURE)) {
+        if (!preg_match_all(self::FLEXIBLE_DOC_STRING_REGEX, $code, $matches, PREG_SET_ORDER|PREG_OFFSET_CAPTURE)) {
             // No heredoc/nowdoc found
-            return null;
+            return $code;
         }
 
         // Keep track of how much we need to adjust string offsets due to the modifications we
@@ -93,19 +150,31 @@ REGEX;
             }
         }
 
-        if (empty($this->patches)) {
-            // We did not end up emulating anything
-            return null;
-        }
-
         return $code;
     }
 
-    private function fixupTokens() {
-        assert(count($this->patches) > 0);
+    private function isEmulationNeeded(string $code): bool
+    {
+        if ($this->isHeredocNowdocEmulationNeeded($code)) {
+            return true;
+        }
+
+        if ($this->isCoalesceEqualEmulationNeeded($code)) {
+            return true;
+        }
+
+        return false;
+    }
+
+    private function fixupTokens()
+    {
+        if (\count($this->patches) === 0) {
+            return;
+        }
 
         // Load first patch
         $patchIdx = 0;
+
         list($patchPos, $patchType, $patchText) = $this->patches[$patchIdx];
 
         // We use a manual loop over the tokens, because we modify the array on the fly